New dtree and dforest modules
