+++ examples/rotational_classifier.rs
@@ -8,8 +8,9 @@ use std::iter::Sum;
 
 use rand::{thread_rng, Rng};
 
-use forester::{DeterministicForestBuilder, DeterministicTreeBuilder};
 use forester::data::{SampleDescription, TrainingData};
+use forester::dforest::DeterministicForestBuilder;
+use forester::dtree::DeterministicTreeBuilder;
 use forester::split::{BestRandomSplit,Split};
 use forester::array_ops::Partition;
 use forester::iter_mean::IterMean;
